fix(renderer): improve raster precision and vanilla compositing
Checkpoint the ongoing comparison against vanilla on Lavapipe. Preserve
lighting and fragment precision through shading, match UNORM blending and
texture filtering, and interpolate sky and cloud fog at the correct stage.
Correct subpixel coverage, GUI preview depth writes, held-item atlas UVs,
and distant depth and cloud coordinates.

The inventory panel matches exactly. The full 854x480 scene still has
9,715 differing pixels, including two with channel errors above two, so
full-scene pixel parity remains in progress. Keep native validation opt-in
and exclude incidental server entities from the deterministic fixture.

+++ mod/src/main/java/com/soulfiremc/server/renderer/ProjectedTriangle.java
@@ -23,5 +23,5 @@ public record ProjectedTriangle(
   ProjectedVertex v1,
   ProjectedVertex v2,
   RenderMaterial material,
-  float sortDepth
+  double sortDepth
 ) {}
